Increment version 0.2.1 to 1.0.0
Need to test this locally before merge and publish
so have set a release date of 09/08/2019 for now.

The API itself currently has a prod environment
on live-1 so this could be released (or as rc-1),
however, the APIs Continous integration is not yet
set up to deploy to live-1 (only live-0). So, should
probably hold off release of this til the live-1
CI deploy is done.