Allow Faraday to be updated
The version restriction on the Faraday gem was `'>= 0.9.2', '< 1.4.0'` and is
now `'~> 1.0', meaning that it can be updated to the latest version 1.x.y
(currently 1.10.0) but version 0.x.y and > 2.0 are excluded.

According to the change log the most significant update of Faraday that this
will allow is version 1.4.
See https://github.com/lostisland/faraday/releases/tag/v1.4.0