Mark test as pending
At some point the Fee Calculator started returning a hash:

```
{
  "case_date": [
    "Enter a valid date."
  ]
}
```

for errors where previously an array was returned. This means that the message
in the exception is different from in the test. Other error responses are still
returned as arrays so it is not possible simply to change
`LAA::FeeCalculator::ResponseError#to_s`.